+++ Prison.App.Web/Controllers/EmployeeController.cs
@@ -30,16 +30,23 @@ namespace Prison.App.Web.Controllers
 
             if (Employees==null)
             {
-               return RedirectToAction("action","controller");
+                return RedirectToAction("Index", "Error");
             }
             return View(Employees);
         }
 
         public ActionResult Details(int id)
         {
-            var Employee = db.GetEmployeeByID(id);
+            if (ArgumentHelper.IsValidID(id))
+            {
+                var Employee = db.GetEmployeeByID(id);
 
-            return View(Employee);
+                return View(Employee);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Error");
+            }
         }
 
         public ActionResult Create()
@@ -60,9 +67,16 @@ namespace Prison.App.Web.Controllers
 
         public ActionResult Edit(int id)
         {
-            var emp = db.GetEmployeeByID(id);
+            if (ArgumentHelper.IsValidID(id))
+            {
+                var emp = db.GetEmployeeByID(id);
 
-            return View(emp);
+                return View(emp);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Error");
+            }
         }
 
         [HttpPost]
@@ -79,17 +93,31 @@ namespace Prison.App.Web.Controllers
 
         public ActionResult Delete(int id)
         {
-            var emp = db.GetEmployeeByID(id);
+            if (ArgumentHelper.IsValidID(id))
+            {
+                var emp = db.GetEmployeeByID(id);
 
-            return View(emp);
+                return View(emp);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Error");
+            }
         }
 
         [HttpPost]
         public ActionResult DeleteFromDb(int id)
         {
-            db.Delete(id);
+            if (ArgumentHelper.IsValidID(id))
+            {
+                db.Delete(id);
 
-            return View("Index", db.GetAllRecordsFromTable());
+                return View("Index", db.GetAllRecordsFromTable());
+            }
+            else
+            {
+                return RedirectToAction("Index", "Error");
+            }
         }
     }
 }
